Small changes to demo drivers to remove compiler warnings
Renamed Problem::dof_pt() to Problem::global_dof_pt() to make explicit that
it returns a pointer to the dof index by the GLOBAL equation number

Modified the HSL frontal solver wrapper so that it issues and error if you 
try to run the solver in parallel.

Fixed a quoting problem in Makefile.am of the build machinary for MUMPS

Tweaked a tolerance in the heat_transfer_and_melting directory



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@702 595e1d82-adf1-4282-bac7-b6073fa2ca8e

+++ demo_drivers/biharmonic/unstructured_2d_biharmonic/unstructured_2d_biharmonic_curvedelement.cc
@@ -740,7 +740,7 @@ fill_in_generic_residual_contribution_biharmonic(Vector<double> &residuals,
    if(bd_element==1)
     {
      // Assign the nodal local equation numbers
-     DenseMatrix<double> nodal_eqn(6,n_position_type,-1);
+     DenseMatrix<int> nodal_eqn(6,n_position_type,-1);
      for(unsigned k=0;k<n_position_type;k++)
       {
        nodal_eqn(0,k) = this->nodal_local_eqn(0,k);
